+++ configurator/CHANGELOG.md
@@ -25,6 +25,19 @@
 - add jump and scale latch pickup modes
 - add takeover modes to manual (#437)
 
+## 1.8.1 (2026-02-24)
+
+### Fixes
+
+- check faderpunk version on landing page (#440)
+
+## 1.8.0 (2026-02-23)
+
+### Features
+
+- add jump and scale latch pickup modes
+- add takeover modes to manual (#437)
+
 ## 1.8.0-beta.0 (2026-02-11)
 
 ### Features